Module aws_sdk_iotjobsdataplane::error

Modules

certificate_validation_exception

See CertificateValidationException

invalid_request_exception

See InvalidRequestException

invalid_state_transition_exception

See InvalidStateTransitionException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

terminal_state_exception

See TerminalStateException

throttling_exception

See ThrottlingException

Structs

CertificateValidationException

The certificate is invalid.

DescribeJobExecutionError

GetPendingJobExecutionsError

InvalidRequestException

The contents of the request were invalid. For example, this code is returned when an UpdateJobExecution request contains invalid status details. The message contains details about the error.

InvalidStateTransitionException

An update attempted to change the job execution to a state that is invalid because of the job execution's

ResourceNotFoundException

The specified resource does not exist.

ServiceUnavailableException

The service is temporarily unavailable.

StartNextPendingJobExecutionError

TerminalStateException

The job is in a terminal state.

ThrottlingException

The rate exceeds the limit.

UpdateJobExecutionError

Enums

DescribeJobExecutionErrorKind

GetPendingJobExecutionsErrorKind

StartNextPendingJobExecutionErrorKind

UpdateJobExecutionErrorKind